Renamed level geometry module
